+++ packages/api/src/bindings.d.ts
@@ -79,6 +79,10 @@ export type NFT = {
   /**
    * Name of the JWT token used to create this NFT.
    */
+  name?: string
+  /**
+   * Optional name of the file(s) uploaded as NFT.
+   */
   scope: string
   /**
    * Date this NFT was created in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format: YYYY-MM-DDTHH:MM:SSZ.