+++ packages/website/pages/stats.js
@@ -7,11 +7,13 @@ import decorateAdditionalCalculatedValues, {
 import { API } from '../lib/api'
 import Loading from '../components/loading'
 import { abbreviateNumber } from 'js-abbreviation-number'
+import Img from '../components/cloudflareImage'
 
 /**
  *
  * @returns {{ props: import('../components/types.js').LayoutProps}}
  */
+
 export function getStaticProps() {
   const logos = fs.readdirSync('public/images/marketplace-logos/home')
   // make opensea be the first logo
@@ -114,9 +116,12 @@ export default function Stats({ logos }) {
           <div className="stat-cards">
             <StatCard title="Upload Count">
               <div>
-                <img
+                <Img
                   src={'/images/stats-upload-count.svg'}
                   alt="Upload Count"
+                  width="500px"
+                  height="200px"
+                  layout="responsive"
                 />
                 <div className="py-4 px-4">
                   <p className="chicagoflf">Total uploads to NFT.Storage</p>